==> R/vcf2genos/0 Load vcf2genos functions.R <==
source.dir <- "/Users/Shared/KKMDocuments/Documents/Github.Repos/amyfrey/VCF_2_genotype/"

source(paste0(source.dir, "R/violin_plot.r"))
source(paste0(source.dir, "R/heat_map.r"))
source(paste0(source.dir, "R/biallelic.r"))
source(paste0(source.dir, "R/create_tgt.r"))
source(paste0(source.dir, "R/create_gtype_df_csv.r"))
source(paste0(source.dir, "R/depth_by_locus.r"))
source(paste0(source.dir, "R/create_gtype_cts_ratios_csv.r"))
source(paste0(source.dir, "R/multiplot_pdf.r"))
source(paste0(source.dir, "R/filter.on.AR.and.minDP.r"))
source(paste0(source.dir, "R/Freebayes.miscalls.r"))
